[UIMA-1431] Fix int computation of average to work with values approaching 2 gb (the largest positive int), without overflowing.  Also add comment to ResultSpecification_impl on why string == test is OK.
git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/uimaj/trunk@793078 13f79535-47bb-0310-9956-ffa450edef68

+++ uimaj-core/src/main/java/org/apache/uima/cas/impl/FSVectorIndex.java
@@ -191,7 +191,7 @@ public class FSVectorIndex extends FSLeafIndexImpl {
     int i; // Current position
     int comp; // Compare value
     while (start <= end) {
-      i = (start + end) / 2;
+      i = (int)(((long)start + end) / 2);
       comp = compare(ele, array[i]);
       if (comp == 0) {
         return i;
